Bugfix: In DreamPair the FixShift function needs a float not an int. Additionally the ME should be filled starting the first bin even if there is no SE counts available.

+++ DreamFunction/DreamPair.h
@@ -65,7 +65,7 @@ class DreamPair {
   ;
   int GetNDists();
   void ShiftForEmpty(DreamDist* pair);
-  void FixShift(DreamDist* pair, DreamDist* otherDist, int kMin);
+  void FixShift(DreamDist* pair, DreamDist* otherDist, float kMin);
   void Rebin(DreamDist* pair, int rebin);
   void ReweightMixedEvent(DreamDist* pair, float kSMin, float kSMax);
   void WriteOutput(TList *Outlist);
